+++ src/consumer/mod.rs
@@ -74,12 +74,23 @@ pub trait ConsumerContext: ClientContext {
         unsafe {
             match err {
                 RDKafkaRespErr::RD_KAFKA_RESP_ERR__ASSIGN_PARTITIONS => {
-                    rdsys::rd_kafka_assign(native_client.ptr(), tpl.ptr());
+                    match native_client.rebalance_protocol() {
+                        RebalanceProtocol::Cooperative => {
+                            rdsys::rd_kafka_incremental_assign(native_client.ptr(), tpl.ptr());
+                        }
+                        _ => {
+                            rdsys::rd_kafka_assign(native_client.ptr(), tpl.ptr());
+                        }
+                    }
                 }
-                _ => {
-                    // Also for RD_KAFKA_RESP_ERR__REVOKE_PARTITIONS
-                    rdsys::rd_kafka_assign(native_client.ptr(), ptr::null());
-                }
+                _ => match native_client.rebalance_protocol() {
+                    RebalanceProtocol::Cooperative => {
+                        rdsys::rd_kafka_incremental_unassign(native_client.ptr(), tpl.ptr());
+                    }
+                    _ => {
+                        rdsys::rd_kafka_assign(native_client.ptr(), ptr::null());
+                    }
+                },
             }
         }
         trace!("Running post-rebalance with {:?}", rebalance);
@@ -166,6 +177,16 @@ unsafe impl KafkaDrop for RDKafkaConsumerGroupMetadata {
 
 unsafe impl Send for ConsumerGroupMetadata {}
 unsafe impl Sync for ConsumerGroupMetadata {}
+
+/// The rebalance protocol for a consumer.
+pub enum RebalanceProtocol {
+    /// The consumer has not (yet) joined a group.
+    None,
+    /// Eager rebalance protocol.
+    Eager,
+    /// Cooperative rebalance protocol.
+    Cooperative,
+}
 
 /// Common trait for all consumers.
 ///
@@ -325,4 +346,7 @@ where
 
     /// Resumes consumption for the provided list of partitions.
     fn resume(&self, partitions: &TopicPartitionList) -> KafkaResult<()>;
+
+    /// Reports the rebalance protocol in use.
+    fn rebalance_protocol(&self) -> RebalanceProtocol;
 }